[FIX+IMP] purchase_discount: report + tests + comments + dates
* Adjusted purchase.report to new inheritance structure
* Split tests in 2 files
* Added comments about PR for using the hook
* Fix date handling
