+++ purchase_discount/tests/__init__.py
@@ -1,3 +1,4 @@
-# (c) 2015 Pedro M. Baeza
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from . import test_product_supplierinfo_discount
 from . import test_purchase_discount
